chore(codegen): refactor forks
