+++ packages/codegen/src/core/modules/JavaScript/JavaScriptCompiler/forks/index.ts
@@ -1,7 +1,7 @@
-import { createPredicates } from './core';
+import { createPredicates } from './serializer';
 
 export const forks = {
-	functions: {
+	serializer: {
 		createPredicates,
 	},
 };
